requirements.txt
streamlit==1.30.0
rembg
Pillow
python-dotenv
